+++ core/src/main/java/org/apache/spark/io/RecordReader.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark.input;
+package org.apache.spark.io;
 
 import java.io.Closeable;
 import java.io.IOException;
